Move code for flushes and merges to a new module
The `Internal` module is growing rather large, this commit moves functionality
related to merge scheduling to a new module. This is in preparation for
implementing properly scheduled merges.
